House Price Predict
The House Price Prediction project aims to create a predictive model using Linear Regression to estimate the prices of houses based on various features and attributes. Linear Regression is a foundational machine learning algorithm that fits a linear relationship between the independent variables (features) and the dependent variable (target) to make predictions.